// src/shared/enums/product.enum.ts
export enum PriceType {
  NORMAL = "NORMAL",
  INTERNET = "INTERNET",
  CMR = "CMR",
}

export enum UnitType {
  UNIT = "UND",
  OTHER = "OTHER",
}

export enum BadgeType {
  DISCOUNT = "DISCOUNT",
  BANK_PROMOTION = "BANK_PROMOTION",
}

export enum Highlights {
  HEIGHT = "ALTO",
  WIDTH = "ANCHO",
  LENGTH = "LARGO",
  MATERIAL = "MATERIAL",
  MODEL = "MODELO",
  TYPE = "TIPO",
  USAGE = "USO",
}

export enum OnImageHover {
  Zoom = "ZOOM",
}

export enum MediaType {
  Sku = "SKU",
}

export enum Symbol {
  COP = "$",
}

export enum SortOrder {
  NAME_ASC = 'name_asc',
  BRAND_ASC = 'brand_asc',
  PRICE_ASC = 'price_asc',
  PRICE_DESC = 'price_desc',
}